+++ src/boxes/account/account_board/TermAndCondition.tsx
