+++ src/boxes/account/account_board/index.ts
@@ -1,0 +1,9 @@
+import Login from './Login';
+import TermAndCondition from './TermAndCondition';
+import Join from './Join';
+
+export {
+  Login,
+  TermAndCondition,
+  Join
+}
