+++ src/boxes/account/index.ts
@@ -1,0 +1,7 @@
+import AccountContainer from './AccountContainer';
+import AccountHeader from './AccountHeader';
+
+export {
+  AccountContainer,
+  AccountHeader
+}
