+++ src/boxes/admin/AdminHeader.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from 'react-router-dom';
-import { Logo } from '../items';
+import { Logo } from '../../items';
 import './AdminPage.css';
 
 const AdminHeader = () => {
