+++ src/boxes/admin/admin_board/index.ts
@@ -1,0 +1,5 @@
+import AdminTest from './AdminTest';
+
+export {
+  AdminTest
+}
