+++ src/boxes/admin/index.ts
@@ -1,0 +1,9 @@
+import AdminHeader from './AdminHeader';
+import AdminMenuBar from './AdminMenuBar';
+import AdminContainer from './AdminContainer';
+
+export {
+  AdminHeader,
+  AdminMenuBar,
+  AdminContainer
+}
