--- src/boxes/board/TestTwo.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const TestTwo = () => {
-  return (
-    <div>
-      테스트용입니다.
-    </div>
-  );
-};
-
-export default TestTwo;
