--- src/boxes/board/index.ts
@@ -1,27 +1,0 @@
-import MerchantList from './MerchantList';
-import FindByMap from './FindByMap';
-import FindBestRoute from './FindBestRoute';
-import Recommendation from './Recommendation';
-import FindByTag from './FindByTag';
-import Notice from './Notice';
-import BuyLocalCurrency from './BuyLocalCurrency';
-import Login from './Login';
-import TermAndCondition from './TermAndCondition';
-import Join from './Join';
-import TestTwo from './TestTwo';
-import AdminTest from './AdminTest';
-
-export {
-    MerchantList,
-    FindByMap,
-    FindBestRoute,
-    Recommendation,
-    FindByTag,
-    Notice,
-    BuyLocalCurrency,
-    Login,
-    TermAndCondition,
-    Join,
-    TestTwo,
-    AdminTest
-}
