+++ src/boxes/common/CommonHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { Logo } from '../items';
+import { Logo } from '../../items';
 import './CommonPage.css'
