+++ src/boxes/common/board/index.ts
@@ -1,0 +1,14 @@
+import { MerchantList, FindByMap, FindBestRoute } from "./map";
+import { Recommendation, FindByTag } from "./recommandation";
+import { Notice } from "./notice";
+import { BuyLocalCurrency } from "./purchase";
+
+export {
+  MerchantList,
+  FindByMap,
+  FindBestRoute,
+  Recommendation,
+  FindByTag,
+  Notice,
+  BuyLocalCurrency,
+};
