+++ src/boxes/common/board/map/FindByMap.tsx
