+++ src/boxes/common/board/map/MerchantList.tsx
