+++ src/boxes/common/board/map/index.ts
@@ -1,0 +1,9 @@
+import FindBestRoute from './FindBestRoute';
+import FindByMap from './FindByMap';
+import MerchantList from './MerchantList';
+
+export {
+  FindBestRoute,
+  FindByMap,
+  MerchantList
+}
