+++ src/boxes/common/board/notice/Notice.tsx
