+++ src/boxes/common/board/notice/index.ts
@@ -1,0 +1,5 @@
+import Notice from './Notice';
+
+export {
+  Notice
+}
