+++ src/boxes/common/board/purchase/index.ts
@@ -1,0 +1,5 @@
+import BuyLocalCurrency from './BuyLocalCurrency';
+
+export {
+  BuyLocalCurrency
+}
