+++ src/boxes/common/board/recommandation/index.ts
@@ -1,0 +1,7 @@
+import FindByTag from './FindByTag';
+import Recommendation from './Recommendation';
+
+export {
+  FindByTag,
+  Recommendation
+}
