+++ src/boxes/common/index.ts
@@ -1,0 +1,9 @@
+import CommonHeader from './CommonHeader';
+import CommonMenuBar from './CommonMenuBar';
+import CommonContainer from './CommonContainer';
+
+export {
+  CommonHeader,
+  CommonMenuBar,
+  CommonContainer
+}
