+++ src/boxes/main/index.ts
@@ -1,0 +1,11 @@
+import MainLogo from './MainLogo';
+import MainNav from './MainNav';
+import MainSearchBar from './MainSearchBar';
+import MainMap from './MainMap';
+
+export {
+  MainLogo,
+  MainNav,
+  MainSearchBar,
+  MainMap
+}
